less compile warnings
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1399 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/src/libafanasy/addresseslist.cpp
@@ -24,7 +24,7 @@ AddressesList::AddressesList( Msg * msg)
 void AddressesList::readwrite( Msg * msg)
 {
    uint32_t size;
-   if( msg->isWriting()) size = addresses.size();
+   if( msg->isWriting()) size = uint32_t( addresses.size());
    rw_uint32_t( size, msg);
    if( size )
    {
